+++ woo-cart-abandonment-recovery/modules/cart-abandonment/classes/class-cartflows-ca-order-table.php
@@ -26,6 +26,14 @@ class Cartflows_Ca_Order_Table extends WP_List_Table {
 	private static $instance;
 
 	/**
+	 * URL of this page
+	 *
+	 * @var   string
+	 * @since 1.2.27
+	 */
+	public $base_url;
+
+	/**
 	 *  Initiator
 	 */
 	public static function get_instance() {
@@ -41,12 +49,26 @@ class Cartflows_Ca_Order_Table extends WP_List_Table {
 	public function __construct() {
 		global $status, $page;
 
+		$this->define_order_table_constants();
+
 		parent::__construct(
 			array(
 				'singular' => 'id',
 				'plural'   => 'ids',
 			)
 		);
+
+		$this->base_url = admin_url( 'admin.php?page=' . WCF_CA_PAGE_NAME . '&action=' . WCF_ACTION_REPORTS );
+	}
+
+	/**
+	 * Define the order table constants.
+	 *
+	 * @since 1.2.27
+	 * @return void
+	 */
+	public function define_order_table_constants() {
+		define( 'WCF_REPORTS_TABLE_ACTION', 'edit_reports_table_actions' );
 	}
 
 	/**
@@ -83,11 +105,39 @@ class Cartflows_Ca_Order_Table extends WP_List_Table {
 
 		$actions = array(
 			'view'   => sprintf( '<a href="%s">%s</a>', esc_url( $view_url ), __( 'View', 'woo-cart-abandonment-recovery' ) ),
-			'delete' => sprintf( '<a onclick="return confirm(\'Are you sure to delete this order?\');" href="?page=%s&action=delete&id=%s">%s</a>', esc_html( $page ), esc_html( $item['id'] ), __( 'Delete', 'woo-cart-abandonment-recovery' ) ),
+			'delete' => sprintf(
+				'<a onclick="return confirm(\'Are you sure to delete this order?\');" href="' . wp_nonce_url(
+					add_query_arg(
+						array(
+							'action' => 'delete',
+							'page'   => esc_html( $page ),
+							'id'     => esc_html( $item['id'] ),
+						),
+						$this->base_url
+					),
+					WCF_REPORTS_TABLE_ACTION,
+					WCF_REPORTS_TABLE_ACTION . '_nonce'
+				) . '">%s</a>',
+				__( 'Delete', 'woo-cart-abandonment-recovery' )
+			),
 		);
 
 		if ( WCF_CART_ABANDONED_ORDER === $item['order_status'] && ! $item['unsubscribed'] ) {
-			$actions['unsubscribe'] = sprintf( '<a onclick="return confirm(\'Are you sure to unsubscribe this user? \');" href="?page=%s&action=unsubscribe&id=%s">%s</a>', esc_html( $page ), esc_html( $item['id'] ), __( 'Unsubscribe', 'woo-cart-abandonment-recovery' ) );
+			$actions['unsubscribe'] = sprintf(
+				'<a onclick="return confirm(\'Are you sure to unsubscribe this user? \');" href="' . wp_nonce_url(
+					add_query_arg(
+						array(
+							'action' => 'unsubscribe',
+							'page'   => esc_html( $page ),
+							'id'     => esc_html( $item['id'] ),
+						),
+						$this->base_url
+					),
+					WCF_REPORTS_TABLE_ACTION,
+					WCF_REPORTS_TABLE_ACTION . '_nonce'
+				) . '">%s</a>',
+				__( 'Unsubscribe', 'woo-cart-abandonment-recovery' )
+			);
 
 		}
 
@@ -276,39 +326,42 @@ class Cartflows_Ca_Order_Table extends WP_List_Table {
 	 */
 	public function process_bulk_action() {
 		global $wpdb;
-		$table_name = $wpdb->prefix . CARTFLOWS_CA_CART_ABANDONMENT_TABLE;
-		$ids        = array();
-		//phpcs:disable WordPress.Security.NonceVerification.Recommended
-		if ( isset( $_REQUEST['id'] ) ) {
 
-			if ( is_array( $_REQUEST['id'] ) ) {
-				$request_id = array_map( 'intval', $_REQUEST['id'] );
-				$ids        = implode( ',', $request_id );
-			} else {
-				$ids = intval( $_REQUEST['id'] );
+		$security_nonce = Cartflows_Ca_Helper::get_instance()->sanitize_text_filter( WCF_REPORTS_TABLE_ACTION . '_nonce', 'GET' );
+
+		// Process the actions only if the nonce is verified and the current user has the capability to manage it.
+		if ( ! empty( $security_nonce ) && wp_verify_nonce( $security_nonce, WCF_REPORTS_TABLE_ACTION ) && current_user_can( 'manage_woocommerce' ) ) {
+
+			$table_name = $wpdb->prefix . CARTFLOWS_CA_CART_ABANDONMENT_TABLE;
+			$ids        = array();
+
+			if ( isset( $_REQUEST['id'] ) ) {
+
+				if ( is_array( $_REQUEST['id'] ) ) {
+					$request_id = array_map( 'intval', $_REQUEST['id'] );
+					$ids        = implode( ',', $request_id );
+				} else {
+					$ids = intval( $_REQUEST['id'] );
+				}
+			}
+
+			if ( ! empty( $ids ) ) {
+				switch ( $this->current_action() ) {
+					case 'delete':
+						// Can't use placeholders for table/column names, it will be wrapped by a single quote (') instead of a backquote (`).
+						$wpdb->query(
+							"DELETE FROM {$table_name} WHERE id IN($ids)" //phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
+						); // db call ok; no cache ok.
+						break;
+					case 'unsubscribe':
+						$wpdb->query(
+							"UPDATE {$table_name} SET unsubscribed = 1 WHERE id IN($ids)" //phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
+						); // db call ok; no cache ok.
+						break;
+
+				}
 			}
 		}
-
-		if ( ! empty( $ids ) ) {
-
-			switch ( $this->current_action() ) {
-
-				case 'delete':
-					// Can't use placeholders for table/column names, it will be wrapped by a single quote (') instead of a backquote (`).
-					$wpdb->query(
-						"DELETE FROM {$table_name} WHERE id IN($ids)" //phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
-					); // db call ok; no cache ok.
-					break;
-				case 'unsubscribe':
-					$wpdb->query(
-						"UPDATE {$table_name} SET unsubscribed = 1 WHERE id IN($ids)" //phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
-					); // db call ok; no cache ok.
-					break;
-
-			}
-		}
-
-		//phpcs:enable WordPress.Security.NonceVerification.Recommended
 	}
